+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - Users are welcomed and able to create their new password to access the service
 - Service name updated from "Overseas" to "Official"
 - Users can create and view Funds; users can create and view Activities
+- Users can download an XML representation of an Activity